added new config for changing the constellation group used (valid combos from 1 to 7. 7 is gps+bdu+glo. 3 is gps+bud. 1 is gps only. Wondering whether glonass is messing up time sync. Also, I hadn't realized millis() is only 32-bit. Fixed that everywhere to make sure duration calc is correct whenever millis rolls over (unsigned so diff always gets accurate duration numbers, as long as only one rolover during the duration measurement. Some other little cleanup things to help debug time synchronizaiton. Forced time update every 1 minute, no matter what, ignored time when the thousandths is not .000 (because of our assumed skew hack..want to only use gps time that's 1 sec aligned)

+++ tracker/cw_functions.cpp
@@ -593,7 +593,7 @@ void cw_send_message() {
     // we send 'PARIS ' 3x above
 
     V1_print(F(EOL "Should start to hear CW now.. RF should be toggling" EOL));
-    uint64_t start_millis = millis();
+    uint32_t start_millis = millis();
     cw_send(cw_msg);
     uint32_t actual_duration = millis() - start_millis;
